requirements.txt
langchain==0.1.20
langchain-openai==0.1.7
langchain-community
langchain-text-splitters
openai==1.30.5
faiss-cpu
python-dotenv